+++ experimental/rule_inference/tests/test_resources/demo/before.java
@@ -1,0 +1,6 @@
+import java.utils.ArrayList;
+class Test {
+    public static void main(String[] args) {
+        data = ContextExecutors.newFixedThreadPool(1);
+    }
+}