+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+Version 0.3.10
+-------------
+* Add support for replace node index
+* Bug fix for iOS cleanup 
+  
+Version 0.3.9
+-------------
+* Rule graph validation #493
+* Bug fix #497 #499
+* Kotlin dependency update
+
 Version 0.3.8
 -------------
 * Bug Fix related to code snippet mode #489